requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argcomplete==1.10.3
beautifulsoup4==4.8.2
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
compressed-rtf==1.0.6
cssselect==1.2.0
distlib==0.3.7
distro==1.8.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
eutils==0.6.0
extract-msg==0.28.7
faiss-cpu==1.7.4
filelock==3.13.1
h11==0.14.0
habanero==1.2.3
httpcore==1.0.2
httpx==0.25.1
humanfriendly==10.0
idna==3.4
IMAPClient==2.1.0
iniconfig==2.0.0
Levenshtein==0.23.0
lxml==4.9.3
metapub==0.5.5
numpy==1.26.2
olefile==0.46
openai==1.3.4
packaging==23.2
pandas==2.1.3
pdfminer.six==20191110
Pillow==10.1.0
platformdirs==4.0.0
pluggy==0.13.1
py==1.11.0
pycryptodome==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
pyproject-api==1.6.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
python-pptx==0.6.23
pytz==2023.3.post1
rapidfuzz==3.5.2
requests==2.31.0
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
striprtf==0.0.26
tabulate==0.9.0
textract==1.6.5
toml==0.10.2
tox==3.4.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
Unidecode==1.3.7
urllib3==2.1.0
virtualenv==20.24.7
xlrd==1.2.0
XlsxWriter==3.1.9
